pyproject.toml
[project]
name = "winners-kdsh-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "centimators[dspy]>=0.3.1",
    "dspy>=3.1.0",
    "pathway[all]>=0.28.0",
    "polars>=1.37.0",
    "python-dotenv>=1.2.1",
]
